searching and paging

+++ ASP_Net/WebApps/MVC/ContosoUniversity/Controllers/StudentsController.cs
@@ -20,9 +20,46 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Students
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder,string currentFilter, string  searchString,int? pageNumber)
         {
-            return View(await _context.Students.ToListAsync());
+            ViewData["CurrentSort"] = sortOrder;
+            ViewData["NameSortParm"]= String.IsNullOrEmpty(sortOrder)? "name_desc":"";
+            ViewData["DateSortParm"]= sortOrder == "Date"? "date_Desc": "Date";
+           
+           if (searchString ==null){
+               pageNumber=1;
+           }
+           else{
+               searchString=currentFilter;
+           }
+           
+           
+            ViewData["CurrentFilter"] = searchString;
+            var students = from s in _context.Students
+                            select s;
+
+            if(!String.IsNullOrEmpty(searchString)){
+                students =students.Where(s => s.LastName.Contains(searchString)
+                                        ||s.FirstMidName.Contains(searchString));
+            }
+
+            switch(sortOrder){
+                case "name_desc":
+                    students = students.OrderByDescending(s =>s.LastName);
+                    break;
+                case "Date":
+                    students= students.OrderBy(s => s.EnrollmentDate);
+                    break;
+                case "date_desc":
+                    students = students.OrderByDescending(s => s.EnrollmentDate);
+                    break;
+                default:
+                    students = students.OrderBy(s => s.LastName);
+                    break;
+            }
+            int pageSize =3;
+            // return View(await students.AsNoTracking().ToListAsync());
+            return View(await PaginatedList<Student>.CreateAsync(students.AsNoTracking(), pageNumber?? 1,pageSize));
         }
 
         // GET: Students/Details/5
@@ -148,7 +185,7 @@ namespace ContosoUniversity.Controllers
         }
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? SaveChangesError = false)
         {
             if (id == null)
             {
@@ -156,10 +193,17 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
+            }
+
+            if (SaveChangesError.GetValueOrDefault ()){
+                ViewData["ErrorMessage"]=
+                "Delete failed.Try  again, and if teh problem persists "+
+                "see your system adminstrator.";
             }
 
             return View(student);
@@ -171,9 +215,16 @@ namespace ContosoUniversity.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
+            if(student == null){
+                return RedirectToAction(nameof(Index));
+            }
+            try{
             _context.Students.Remove(student);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+            }catch(DbUpdateException){
+                return RedirectToAction(nameof(Delete), new {id=id,SaveChangesError = true});
+            }
         }
 
         private bool StudentExists(int id)
